LayoutTests/http/tests/site-isolation/mouse-events/iframes-with-tooltips-expected.txt

Verifies that tooltips from iframes are correct.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS tooltips[0] is 'Test Main Frame'
PASS tooltips[2] is 'Tooltip A'
PASS tooltips[4] is 'Tooltip B'
PASS successfullyParsed is true

TEST COMPLETE
